Rename provider `virtualworkspace` to `apiexport` and add some docs
On-behalf-of: SAP <marvin.beckers@sap.com>
