+++ examples/apiexport/manifests/workspace.yaml
@@ -1,0 +1,24 @@
+# Copyright 2025 The KCP Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# 
+# This file allows creating further workspaces from the example bundle.
+
+apiVersion: tenancy.kcp.io/v1alpha1
+kind: Workspace
+metadata:
+  generateName: apiexport-
+spec:
+  type:
+    name: examples-apiexport-multicluster
+    path: root
